Modified Get-JiraUser to return user group information.
Group info needs to be returned from Get-JiraUser to provide more flexibility and information for managing JIRA users.

Unfortunately, this requires a second REST call to JIRA, since the user search resource does not provide group information. In testing, I haven't seen a large impact on performance, but this is something to keep in mind.