Removed FunctionsToExport and added Export-ModuleMember to .psm1
Using the FunctionsToExport attribute in the module manifest (PSJira.psd1) was compromising one of the benefits of keeping each function in its own .ps1 file - the highly modular approach that allowed me to add, move, and remove files without modifying the manifest at all.  I still need to specify which functions to export, though, since I don't want the Internal folder being exported.

This approach will dynamically export only the functions in the Functions folder.

I'm a bit worried about performance - this may cause a bit of a performance hit to the initial import of the module - but I'd like to try it and see how it plays out.