Fixed WhatIf, Force, and Confirm behavior
The previous implementation of WhatIf, Force, and Confirm were handling Force and Confirm correctly, and were not performing operations when the user passed -WhatIf, but they were not displaying any WhatIf messages.  Instead of manually checking for $WhatIfPreference, I'm allowing PowerShell to do the work instead, and temporarily overriding the $ConfirmPreference flag if the -Force parameter is set. This behavior produces the correct behavior in all three cases (-WhatIf, no modifiers, and -Force).