Added skip_tags back to appveyor.yml
The only real use I could see for building when I tag a repo would be to rebuild and publish to PSGallery when I create a GitHub release. This only happens when I tag it on GitHub, though - if I tag the commit locally and push, it doesn't trigger a build. It looks like it will be easier all around just to use the release label in my commit message, then manually create a release in GitHub.