Reversed condition order.
Since commits should go primarily against the dev branch now, it makes more sense to check the branch first, then check the commit message.  In the future, it may be appropriate to remove the requirement for the commit message altogether, but for now, we'll stick with this and see how it plays out.