Removed fallback logic for Reporter parameter.
Issue #64 suggests that there are cases where JIRA does not require a reporter for an issue. To accomodate this, we need the option not to provide a reporter.

In the case where a reporter is required and the user doesn't specify one, a new and specific error message is provided (so users don't try to add a reporter object to the Fields parameter).