Move ConvertFrom-Json2 test to Tests dir
This is the start of a re-organization and cleanup of the tests in the module. As the module is becoming pretty large, it may be appropriate to remove the tests from the published module in the PSGallery.

The new Shared.ps1 file can also serve as a helper script with a couple of useful functions. The defProp() function used in most of the tests is a prime example of this.