Additional test cleanup
AppVeyor is causing a lot of these tests to fail, even though they pass on my local machine. They all seem to be failing on tests which check the PSTypeName of the output object.

I'm cleaning them up and generalizing them a bit more, moving shared logic into Shared.ps1, so that if there's an issue with that logic in the tests, it can be fixed once for all tests.