Updated VSCode settings and CI pipeline
As it turns out, there are quite a few cool new features in VSCode and the PowerShell extension. These changes provide better integration with that editor and an improved CI pipeline and framework.

tasks.json has updated the task engien version from 0.1.0 to 2.0.0, allowing VS Code to use the new integrated terminal for tasks instead of the output window.

A new build script (taken from a Plaster example template) provides much greater flexibility in the build process. In theory, we can now build the module locally as well as in AppVeyor (though the environment may still be slightly different).

By implementing the BuildHelpers module by RamblingCookieMonster, we also gain the ability to fill out the FunctionsToExport key in the module manifest automatically on build. This means that the file in the repo no longer needs to be updated for each function, and the version in the gallery takes advantage of the improved load times by specifying each function.

This is still a bit experimental, since I want to see how AppVeyor handles it, but locally everything is working great.