Move Pester result upload to Appveyor.ps1
PSake was causing the job to fail before it reached the AfterTest task where I'd been trying to upload task results. Moving this back to Appveyor.ps1 should cause it to run correctly even when psake exits with an error.

This commit also removes the changes I'd made to build.psake.ps1 (changing Test to CoreTest to enable a BeforeTest and AfterTest task in build.settings.ps1). Those tasks are probably not necessary, and I'd like to keep build.psake.ps1 as close to the template as possible.